refactor(admin): consolidate data explorer module
The storage/data explorer list page, its detail view, query logic and styles were separated across generic folders. Move them under admin/data-explorer so browsing stored data is a single cohesive module.
