+++ webui/src/admin/data-explorer/data-explorer.scss
